+++ AsyncDisplayKit/Private/ASStackUnpositionedLayout.mm
@@ -10,6 +10,7 @@
 
 #import "ASStackUnpositionedLayout.h"
 
+#import <tgmath.h>
 #import <numeric>
 
 #import "ASLayoutSpecUtilities.h"
@@ -87,7 +88,7 @@ static void stretchChildrenAlongCrossDimension(std::vector<ASStackUnpositionedIt
     // restretch all stretchable children along the cross axis using the new min. set their max size to childCrossMax,
     // not crossMax, so that if any of them would choose a larger size just because the min size increased (weird!)
     // they are forced to choose the same width as all the other children.
-    if (alignItems == ASStackLayoutAlignItemsStretch && fabs(cross - childCrossMax) > 0.01) {
+    if (alignItems == ASStackLayoutAlignItemsStretch && std::fabs(cross - childCrossMax) > 0.01) {
       l.layout = crossChildLayout(child, style, stack, stack, childCrossMax, childCrossMax);
     }
   }
@@ -182,7 +183,7 @@ static const CGFloat kViolationEpsilon = 0.01;
  */
 static std::function<BOOL(const ASStackUnpositionedItem &)> isFlexibleInViolationDirection(const CGFloat violation)
 {
-  if (fabs(violation) < kViolationEpsilon) {
+  if (std::fabs(violation) < kViolationEpsilon) {
     return [](const ASStackUnpositionedItem &l) { return NO; };
   } else if (violation > 0) {
     return [](const ASStackUnpositionedItem &l) { return l.child.flexGrow; };
@@ -263,7 +264,7 @@ static void flexChildrenAlongStackDimension(std::vector<ASStackUnpositionedItem>
   }
 
   // Each flexible child along the direction of the violation is expanded or contracted equally
-  const CGFloat violationPerFlexChild = floorf(violation / flexibleChildren);
+  const CGFloat violationPerFlexChild = std::floor(violation / flexibleChildren);
   // If the floor operation above left a remainder we may have a remainder after deducting the adjustments from all the
   // contributions of the flexible children.
   const CGFloat violationRemainder = violation - (violationPerFlexChild * flexibleChildren);
